refactor: add 'task' and 'project' dirs to 'store'
- move reducers, actions, constants and types there;
