+++ src/store/reducers.ts
@@ -1,7 +1,7 @@
 import { combineReducers } from 'redux';
 
-import projectsReducer from 'Store/reducers/projects-reducers';
-import tasksReducer from 'Store/reducers/tasks-reducers';
+import projectsReducer from 'Store/project/projects-reducers';
+import tasksReducer from 'Store/task/tasks-reducers';
 
 const rootReducer = combineReducers({
   tasks: tasksReducer,
